Move merged series spec to integration
